Added one new type and underlying structure
This has an abstract class for Token which other token types can extend.
The mandatory isValid method that everyone must implement is what we are after
so that clients don't need to know what type of token they have to validate it.

This should also simplify token creation. Clients will no longer need to interact
with the "raw" token to know how to set claims etc. They just need to construct
the token type they need and then call encode to pass the token on.