Simplified validation api
Now the only thing needed to validate a JWT source token is to call the new
static Token.validate method, which when it succeeds will return a parsed/
hydrated token type. This combines parsing and validation, and makes the calling
code simpler.

Also streamlined exception handling to guarantee that all validation code will
produce a TokenValidationException when things go wrong.