Fixes broken tests
This fixes the tests that were broken after the last commit. The reason the tests
were failing was that they were not hitting the mocks but rather were hitting the cache.
Adding the header to clear the cache to the first test forced them to run against
the endpoints, allowing for the intended Connection refused message. This won't happen
if the tests hit the cache since the cache doesn't use X-Okapi-Url, but rather fetches
from in memory storage.

The reason these tests were passing in prior commits was that a 400 was encountered
but for different reasons (the finalUserId was null) causing its own 400 error, which
made these tests pass, but for the wrong reasons.

I haven't had a chance to fully explore why these tests when running on the current master
branch don't hit the cache. My current hypothesis is that on master they take slightly
longer to run, allowing for the cache timeout to occur, thus clearing the cache before
they run. This should be something I can verify but as of yet haven't had a chance to do so.