Removing exp claim form AccessToken
Noticed that we were adding this claim when really, for now we shouldn't be
since no action is taken on that claim. This now makes it such that the only
difference between existing tokens and the 'new' tokens is the type claim
which is used extensively by the new code to determine what to do when a given
token arrives for validation.