Tenant init and postInit is working
Working means that all existing unit tests are passing as well as a new unit
test for initializing the tenant which involves creating the table for the tenant
etc.

There are some caveats:
* The order of adding the RouterCreator instances is very important. AuthorizeApi has
to be added last for it not to hijack all requests with the filter. Not sure how secure
this will be ultimately because I believe it means at least one route isn't subject
to authtoken's filter regime.
* Less of a big deal, but it looks like all the tests will need to go in AuthTokenTest
since I wasn't able to spin up a second MainVerticle, which sort of makes sense if
tests are not seqential, which on reflection they are not. The idea was to spin
up a second test class for the token store, but that probably is just cosmetic.
* This has defined the openapi resources which I don't think we will ultimately
want here since mod-authtoken will want to handle all routes (we're not going to
be able to spin up other api classes since everyone should be subject to its filters
I believe). So my next task to to try to remove that material to see if I can still
get it to do postInit without a problem.