Better connection handling
Didn't like how I was creating the connection and closing it outside of
the new class methods. They should handle that. But then got mired in
consistently closing the connection. TNoticed that the vertx-pg-client
provides the withConnection method, which greatly improved things
since it takes care of closing the connectionn as long as a future
is returned.
: