MODAT-110: Implement token persistent store (#113)
Adds persistent storage to mod-authtoken through folio-vertx-lib. Two token types are now persisted: refresh tokens and API tokens. Persistence of these two token types is needed in order to support revoking the tokens under certain conditions. In the case of refresh tokens, tokens are revoked when they are used more than once (this is considered a leaked and compromised token). In the case of API tokens, tokens are revoked when set as revoked.

This work is part of a group of related tickets to improve FOLIO security. Other tickets that are part of this work are MODAT-110, MODAT-111 and MODAT-112.

Integrating folio-vertx-lib (to support storage) has moved much of the code that was in MainVerticle.java into a new class called AuthorizeApi.java.