Wired up a schema to token route
It's working in that the request is aware of the schema and is doing some validation
but one bummer right now is that the 'required' property of the schema isn't doing its
job. Adding it to the schema makes openapi complain that items aren't there when
they are. Perhaps it is picking up the filter request which doesn't have the body?