MODAT-128: jwt.signing.key hint on BadSignatureException
As a sysop I would like to read some hint in the log how to fix the configuration when "Invalid token" might be caused by a wrong jwt.signing.key.

Use case 1:

Given that jwt.signing.key is not set and mod-authtoken is restarted

When using a token that has been issued before the restart that causes an "Invalid token" error

Then the log should indicate this it might be caused by the mod-authtoken restart.

Use case 2:

Given that a cluster of multiple instances of mod-authtoken has been configured with different jwt.signing.key values.

When using a token that has been issued by an instance with different key and getting an "Invalid token" error.

Then the log should indicate this it might be caused by different keys.