Changes to readme and OpenApi yaml
Updated readme to refer to the new OpenApi yaml since before the only documentation
of the endpoints were in the Readme. Now we have the yaml. Clarified the yaml
to include the token expirations.

Also noticed that the tokenResponse.json wasn't being referred to properly in the
yaml. Fixed that.

Finally noticed that the tokenResponse.json didn't have the new expiration properties
in it. Fixed that as well.